+++ angel_container/lib/angel_container.dart
@@ -5,3 +5,4 @@ export 'src/empty/empty.dart';
 export 'src/static/static.dart';
 export 'src/exception.dart';
 export 'src/reflector.dart';
+export 'src/throwing.dart';